Use the dependency log as pinned versions.
All group/module/artifact/version combinations from the dependency log are
used as pinned versions. This means the repositories are not scanned for
better artifacts (i.e. artifacts which produce a match for the constraints
with a higher score). The cache is used in this case without checking the
"rescan" attribute - as no scan is necessary.