Replace the finalizer with actions.
Jonchki v0.0.8.1 and previous have "post triggers". An installer
script can use the method "register_post_trigger" of the install
helper to register a function, which is called after the install
scripts of all artifacts. The "level" parameter offers a simple
way to control the execution order of the scripts.

This patch replaces the finalizer with post triggers. The typical
tasks of a finalizer is to install additional files and to create
an archive. The 2 tasks are now 2 separate post triggers to allow
better control.

To reflect the new use of the post triggers, they are renamed to
"actions".